Add files via upload
This was an intense project to figure out. I had to understand not only splitting elements, checking for lengths of lists, and so much more, but how to problem solve. I used a lot of resources in getting here. Special mention to Geeksforgeeks, Freecodecamp's forums, and Google(so much Google...).